FogBugz #93621: Upgrade socfpga to v3.7
Biggest change in v3.7 is the usage of ARM multiplatform.

Because of multiplatform, the usage of Makefile.boot is gone and the
config AUTO_ZRELADDR is used. Since I don't have a 100% grasp of
a this concept, to make sure that your uImage has the correct entry
address, you need to run this:

'make uImage LOADADDR=0x8000'

Another side affect of v3.7 multiplatform is that this config option:
CONFIG_ARCH_VEXPRESS_CORTEX_A5_A9_ERRATA gets enabled when
socfpga_defconfig is used. This is perfectly acceptable on HW, but if
you would like to use VT, then CONFIG_ARCH_VEXPRESS_CORTEX_A5_A9_ERRATA
along with CONFIG_ARM_ERRATA_751472 needs to be disabled.

Also new in v3.7 is the conversion for platform SMP funtions to
use a unified smp_operations function.

A change in /arm/include/asm/io.h is generating alot of warnings from
the usage of __raw_writel() and __raw_readl() in the DWC OTG driver.
A fix for this will be done when the DWC OTG driver is rewritten.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>