FogBugz #113756: Fix error for FPGA LED PIO base address
The base address for the LED PIO is incorrect. The current address is for
the LWH2F bridge and 0xff2100040 is the address for the LED PIO slave.
The current implementation works because Qsys selected the LED PIO as the
default slave, so all reads/writes from the LWH2F bridge master targeting
holes in the master address map gets routed to the LED PIO anywa.

However, if any peripheral with an address range larger than the PIO's gets
added to the system, then the build malfunctions, since that peripheral
becomes the new default slave.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>