FogBugz #167813-2: Update the USB DTS entry to use the s3c-hsotg driver
Add a "usb-nop-xceiv" PHY node that is needed to be used with the new
generic PHY driver that is from:

[1790d80: drivers: phy: add generic PHY framework]

Add a clock node for dwc2 usb dts entry, as the s3c-hsotg driver will need
to properly initialize. Remove dts bindings that were used by the old
dwc2 driver.

To enable USB peripheral mode on SOCFPGA, build the kernel with the following
defconfig:

CONFIG_USB_S3C_HSOTG=y

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>